Separate keywords into setup and checks.
