+++ git-keeper-robot/gkeeprobot/keywords/ServerSetupKeywords.py
@@ -4,7 +4,7 @@ from gkeeprobot.control.ServerControl import ServerControl
 control = ServerControl()
 
 
-class ServerKeywords:
+class ServerSetupKeywords:
 
     def set_faculty_csv(self, *faculty_list):
         for faculty_name in faculty_list:
@@ -19,26 +19,6 @@ class ServerKeywords:
 
     def stop_gkeepd(self):
         control.run('keeper', 'screen -S gkserver -X quit')
-
-    def server_running(self):
-        result = control.run_server_script('keeper', 'server_running.py')
-        assert result == 'Yes'
-
-    def server_not_running(self):
-        result = control.run_server_script('keeper', 'server_running.py')
-        assert result == 'No'
-
-    def expect_email(self, to_user, contains):
-        result = control.run_server_script('keeper', 'email_check.py', to_user, contains)
-        assert result == 'True'
-
-    def user_exists(self, username):
-        result = control.run_server_script('keeper', 'user_exists.py', username)
-        assert result == 'Yes'
-
-    def user_does_not_exist(self, username):
-        result = control.run_server_script('keeper', 'user_exists.py', username)
-        assert result == 'No'
 
     def remove_user(self, username):
         control.run('keeper', 'sudo userdel -r {} || :'.format(username))
